# FrontEnd1/tools/swagger/src/swagger.json
{
  "swagger" : "2.0",
  "info" : {
    "description" : "Swagger Api documentation for Gestion d'analyse medical  application",
    "version" : "0.0.1-SNAPSHOT",
    "title" : "Swagger Api Documentation for Gestion d'analyse medical application",
    "contact" : {
      "name" : "khalid Abba",
      "url" : "https://www.udemy.com/user/",
      "email" : "khalidoabba1996@gmail.com"
    }
  },
  "basePath" : "http://localhost:8080",
...
  "paths" : {
...
    "/testCovid/v1/absence/{idAbsence}" : {
      "get" : {
        "tags" : [ "testCovidv1absence" ],
        "summary" : "rechercher un absence par ID",
        "description" : "Cette methode permet de rechercher  un absence par son ID ",
        "operationId" : "findById",
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "idAbsence",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int32"
        } ],
        "responses" : {
          "200" : {
            "description" : "absence a ete trouver dans la BDD",
            "schema" : {
              "$ref" : "#/definitions/AbsenceDTO"
            }
          },
          "404" : {
            "description" : "Aucun absence n'existe  dans la BDD avec l'ID fourni"
          }
        }
      }
    },
...
    "/testCovid/v1/analyseMedical/delete/{idAnalyseMedical}" : {
      "delete" : {
        "tags" : [ "analyseMedical" ],
        "summary" : "supprimer  l'analyse Medical ",
        "description" : "Cette methode permet de supprimer l'analyse Medical par ID ",
        "operationId" : "delecte",
        "parameters" : [ {
          "name" : "idAnalyseMedical",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int32"
        } ],
        "responses" : {
          "200" : {
            "description" : "analyse Medical a ete supprimer",
            "schema" : {
              "$ref" : "#/definitions/AnalyseMedicalDTO"
            }
          }
        }
      }
    },
    "/testCovid/v1/analyseMedical/filter{idAnalyseMedical}" : {
      "get" : {
        "tags" : [ "analyseMedical" ],
        "summary" : "rechercher un analyse medical par ID",
        "description" : "Cette methode permet de rechercher  d'analyse medical par son ID ",
        "operationId" : "findById",
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "idAnalyseMedical",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int32"
        } ],
        "responses" : {
          "200" : {
            "description" : "l'analyse medical a ete trouver dans la BDD",
            "schema" : {
              "$ref" : "#/definitions/AnalyseMedicalDTO"
            }
          },
          "404" : {
            "description" : "Aucun analyse medical  n'existe  dans la BDD avec l'ID fourni"
          }
        }
      }
    },
...
    "/testCovid/v1/article/delete/{idArticle}" : {
      "delete" : {
        "tags" : [ "article" ],
        "summary" : "supprimer  un article ",
        "description" : "Cette methode permet de supprimer un article par ID ",
        "operationId" : "delecte",
        "parameters" : [ {
          "name" : "idArticle",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int32"
        } ],
        "responses" : {
          "200" : {
            "description" : "article a ete supprimer",
            "schema" : {
              "$ref" : "#/definitions/ArticleDTO"
            }
          }
        }
      }
    },
    "/testCovid/v1/article/{idArticle}" : {
      "get" : {
        "tags" : [ "article" ],
        "summary" : "rechercher un article par ID",
        "description" : "Cette methode permet de rechercher  un article par son ID ",
        "operationId" : "findById",
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "idArticle",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int32"
        } ],
        "responses" : {
          "200" : {
            "description" : "article a ete trouver dans la BDD",
            "schema" : {
              "$ref" : "#/definitions/ArticleDTO"
            }
          },
          "404" : {
            "description" : "Aucun article n'existe  dans la BDD avec l'ID fourni"
          }
        }
      }
    },
...
    "/testCovid/v1/bilanfinance/delete/{idBilan}" : {
      "delete" : {
        "tags" : [ "testCovidv1bilanfinance" ],
        "summary" : "supprimer  un bilanfinance ",
        "description" : "Cette methode permet de supprimer un bilanfinance par ID ",
        "operationId" : "delecte",
        "parameters" : [ {
          "name" : "idBilan",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int32"
        } ],
        "responses" : {
          "200" : {
            "description" : "bilanfinance a ete supprimer",
            "schema" : {
              "$ref" : "#/definitions/BilanFinancierDTO"
            }
          }
        }
      }
    },
    "/testCovid/v1/bilanfinance/{bilanfinance}" : {
      "get" : {
        "tags" : [ "testCovidv1bilanfinance" ],
        "summary" : "rechercher un bilanfinance par ID",
        "description" : "Cette methode permet de rechercher  un bilanfinance par son ID ",
        "operationId" : "findById",
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "idBilan",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int32"
        } ],
        "responses" : {
          "200" : {
            "description" : " bilanfinance a ete trouver dans la BDD",
            "schema" : {
              "$ref" : "#/definitions/BilanFinancierDTO"
            }
          },
          "404" : {
            "description" : "Aucun bilanfinance n'existe  dans la BDD avec l'ID fourni"
          }
        }
      }
    },
...
    "/testCovid/v1/biologiste/delete/{idBiologiste}" : {
      "delete" : {
        "tags" : [ "biologiste" ],
        "summary" : "supprimer  un biologiste ",
        "description" : "Cette methode permet de supprimer un biologiste par ID ",
        "operationId" : "delecte",
        "parameters" : [ {
          "name" : "idBiologiste",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int32"
        } ],
        "responses" : {
          "200" : {
            "description" : "biologiste a ete supprimer",
            "schema" : {
              "$ref" : "#/definitions/BiologisteDTO"
            }
          }
        }
      }
    },
...
    "/testCovid/v1/biologiste/{idBiologiste}" : {
      "get" : {
        "tags" : [ "biologiste" ],
        "summary" : "rechercher un biologiste par ID",
        "description" : "Cette methode permet de rechercher  un biologiste par son ID ",
        "operationId" : "findById",
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "idBiologiste",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int32"
        } ],
        "responses" : {
          "200" : {
            "description" : "la biologiste a ete trouver dans la BDD",
            "schema" : {
              "$ref" : "#/definitions/BiologisteDTO"
            }
          },
          "404" : {
            "description" : "Aucun biologiste n'existe  dans la BDD avec l'ID fourni"
          }
        }
      }
    },
...
    "/testCovid/v1/boncommande/delete/{idBonCommande}" : {
      "delete" : {
        "tags" : [ "boncommande" ],
        "summary" : "supprimer  boncommande ",
        "description" : "Cette methode permet de supprimer boncommande par ID ",
        "operationId" : "delecte",
        "parameters" : [ {
          "name" : "idBonCommande",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int32"
        } ],
        "responses" : {
          "200" : {
            "description" : "boncommandea ete supprimer",
            "schema" : {
              "$ref" : "#/definitions/BonCommandeDTO"
            }
          }
        }
      }
    },
    "/testCovid/v1/boncommande/filter{idBonCommande}" : {
      "get" : {
        "tags" : [ "boncommande" ],
        "summary" : "rechercher boncommande par ID",
        "description" : "Cette methode permet de rechercher  boncommande par son ID ",
        "operationId" : "findById",
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "idBonCommande",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int32"
        } ],
        "responses" : {
          "200" : {
            "description" : "boncommandea ete trouver dans la BDD",
            "schema" : {
              "$ref" : "#/definitions/BonCommandeDTO"
            }
          },
          "404" : {
            "description" : "boncommande medical  n'existe  dans la BDD avec l'ID fourni"
          }
        }
      }
    },
...
    "/testCovid/v1/compte/{idCompte}" : {
      "get" : {
        "tags" : [ "testCovidv1compte" ],
        "summary" : "rechercher un compte par ID",
        "description" : "Cette methode permet de rechercher  un compte par son ID ",
        "operationId" : "findById",
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "idCompte",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int32"
        } ],
        "responses" : {
          "200" : {
            "description" : "compte a ete trouver dans la BDD",
            "schema" : {
              "$ref" : "#/definitions/CompteDTO"
            }
          },
          "404" : {
            "description" : "Aucun compte n'existe  dans la BDD avec l'ID fourni"
          }
        }
      }
    },
...
    "/testCovid/v1/conge/delete/{idConge}" : {
      "delete" : {
        "tags" : [ "testCovidv1conge" ],
        "summary" : "supprimer  un conge ",
        "description" : "Cette methode permet de supprimer un conge par ID ",
        "operationId" : "delecte",
        "parameters" : [ {
          "name" : "idConge",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int32"
        } ],
        "responses" : {
          "200" : {
            "description" : "patient a ete supprimer",
            "schema" : {
              "$ref" : "#/definitions/CongeDTO"
            }
          }
        }
      }
    },
    "/testCovid/v1/conge/{idConge}" : {
      "get" : {
        "tags" : [ "testCovidv1conge" ],
        "summary" : "rechercher un conge par ID",
        "description" : "Cette methode permet de rechercher  un conge par son ID ",
        "operationId" : "findById",
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "idConge",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int32"
        } ],
        "responses" : {
          "200" : {
            "description" : "conge a ete trouver dans la BDD",
            "schema" : {
              "$ref" : "#/definitions/CongeDTO"
            }
          },
          "404" : {
            "description" : "Aucun conge n'existe  dans la BDD avec l'ID fourni"
          }
        }
      }
    },
...
    "/testCovid/v1/departement/{idDepartement}" : {
      "get" : {
        "tags" : [ "departement" ],
        "summary" : "rechercher un departement par ID",
        "description" : "Cette methode permet de rechercher  un departement par son ID ",
        "operationId" : "findById",
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "idDepartement",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int32"
        } ],
        "responses" : {
          "200" : {
            "description" : "la departement a ete trouver dans la BDD",
            "schema" : {
              "$ref" : "#/definitions/DepartementDTO"
            }
          },
          "404" : {
            "description" : "Aucun departement n'existe  dans la BDD avec l'ID fourni"
          }
        }
      }
    },
...
    "/testCovid/v1/departements/delete/{idDepartement}" : {
      "delete" : {
        "tags" : [ "departement" ],
        "summary" : "supprimer un departement ",
        "description" : "Cette methode permet de supprimer un departement par ID ",
        "operationId" : "delecte",
        "parameters" : [ {
          "name" : "idDepartement",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int32"
        } ],
        "responses" : {
          "200" : {
            "description" : "departement a ete supprimer",
            "schema" : {
              "$ref" : "#/definitions/DepartementDTO"
            }
          }
        }
      }
    },
...
    "/testCovid/v1/detailResultat/delete/{idDetailResultat}" : {
      "delete" : {
        "tags" : [ "detailResultat" ],
        "summary" : "supprimer  un detail Resultat ",
        "description" : "Cette methode permet de supprimer un detail Resultat par ID ",
        "operationId" : "delecte",
        "parameters" : [ {
          "name" : "idDetailResultat",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int32"
        } ],
        "responses" : {
          "200" : {
            "description" : "detail Resultat a ete supprimer",
            "schema" : {
              "$ref" : "#/definitions/DetailResultatDTO"
            }
          }
        }
      }
    },
    "/testCovid/v1/detailResultat/{idDetailResultat}" : {
      "get" : {
        "tags" : [ "detailResultat" ],
        "summary" : "rechercher un detail Resultat par ID",
        "description" : "Cette methode permet de rechercher  un detail Resultat par son ID ",
        "operationId" : "findById",
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "idDetailResultat",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int32"
        } ],
        "responses" : {
          "200" : {
            "description" : "la detail Resultat a ete trouver dans la BDD",
            "schema" : {
              "$ref" : "#/definitions/DetailResultatDTO"
            }
          },
          "404" : {
            "description" : "Aucun detail Resultat n'existe  dans la BDD avec l'ID fourni"
          }
        }
      }
    },
...
    "/testCovid/v1/employe/{idEmploye}" : {
      "get" : {
        "tags" : [ "testCovidv1employe" ],
        "summary" : "rechercher un employe par ID",
        "description" : "Cette methode permet de rechercher  un employe par son ID ",
        "operationId" : "findById",
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "idEmploye",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int32"
        } ],
        "responses" : {
          "200" : {
            "description" : "la employe a ete trouver dans la BDD",
            "schema" : {
              "$ref" : "#/definitions/EmployeDTO"
            }
          },
          "404" : {
            "description" : "Aucun employe n'existe  dans la BDD avec l'ID fourni"
          }
        }
      }
    },
...
    "/testCovid/v1/facture/{idFacture}" : {
      "get" : {
        "tags" : [ "facture" ],
        "summary" : "rechercher d'une facture par ID",
        "description" : "Cette methode permet de rechercher d'une facture par son ID ",
        "operationId" : "findById",
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "idFacture",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int32"
        } ],
        "responses" : {
          "200" : {
            "description" : "la facture a ete trouver dans la BDD",
            "schema" : {
              "$ref" : "#/definitions/FactureDTO"
            }
          },
          "404" : {
            "description" : "Aucun facture n'existe  dans la BDD avec l'ID fourni"
          }
        }
      }
    },
...
    "/testCovid/v1/fournisseur/{idFournisseur}" : {
      "get" : {
        "tags" : [ "fournisseur" ],
        "summary" : "rechercher d'une fournisseur par ID",
        "description" : "Cette methode permet de rechercher d'une fournisseur par son ID ",
        "operationId" : "findById",
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "idFournisseur",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int32"
        } ],
        "responses" : {
          "200" : {
            "description" : "la fournisseur a ete trouver dans la BDD",
            "schema" : {
              "$ref" : "#/definitions/FournisseurDTO"
            }
          },
          "404" : {
            "description" : "Aucun fournisseur n'existe  dans la BDD avec l'ID fourni"
          }
        }
      }
    },
...
    "/testCovid/v1/labo/{idLabo}" : {
      "get" : {
        "tags" : [ "testCovidv1labo" ],
        "summary" : "rechercher un labo par ID",
        "description" : "Cette methode permet de rechercher  un labo par son ID ",
        "operationId" : "findById",
        "produces" : [ "application/json" ],
        "responses" : {
          "200" : {
            "description" : " labo a ete trouver dans la BDD",
            "schema" : {
              "$ref" : "#/definitions/LaboDTO"
            }
          },
          "404" : {
            "description" : "Aucun labo n'existe  dans la BDD avec l'ID fourni"
          }
        }
      }
    },
...
    "/testCovid/v1/mouvement/{idMouvement}" : {
      "get" : {
        "tags" : [ "testCovidv1mouvement" ],
        "summary" : "rechercher un mouvement par ID",
        "description" : "Cette methode permet de rechercher  un mouvement par son ID ",
        "operationId" : "findById",
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "idMouvement",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int32"
        } ],
        "responses" : {
          "200" : {
            "description" : "la patient a ete trouver dans la BDD",
            "schema" : {
              "$ref" : "#/definitions/MouvementDTO"
            }
          },
          "404" : {
            "description" : "Aucun patient n'existe  dans la BDD avec l'ID fourni"
          }
        }
      }
    },
...
    "/testCovid/v1/operation/delete/{idOperation}" : {
      "delete" : {
        "tags" : [ "testCovidv1operation" ],
        "summary" : "supprimer  un operation ",
        "description" : "Cette methode permet de supprimer un operation par ID ",
        "operationId" : "delecte",
        "parameters" : [ {
          "name" : "idOperation",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int32"
        } ],
        "responses" : {
          "200" : {
            "description" : "operation a ete supprimer",
            "schema" : {
              "$ref" : "#/definitions/OperationDTO"
            }
          }
        }
      }
    },
    "/testCovid/v1/operation/filter{dateOperation}" : {
      "get" : {
        "tags" : [ "testCovidv1operation" ],
        "summary" : "rechercher un operation par date",
        "description" : "Cette methode permet de rechercher un operation par son date ",
        "operationId" : "findByDateOperation",
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "dateOperation",
          "in" : "path",
          "required" : true,
          "type" : "string",
          "format" : "date-time"
        } ],
        "responses" : {
          "200" : {
            "description" : "la operation a ete trouver dans la BDD",
            "schema" : {
              "$ref" : "#/definitions/OperationDTO"
            }
          },
          "404" : {
            "description" : "Aucun operation n'existe  dans la BDD avec le date fourni"
          }
        }
      }
    },
    "/testCovid/v1/operation/{idOperation}" : {
      "get" : {
        "tags" : [ "testCovidv1operation" ],
        "summary" : "rechercher un operation par ID",
        "description" : "Cette methode permet de rechercher  un operation par son ID ",
        "operationId" : "findById",
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "idOperation",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int32"
        } ],
        "responses" : {
          "200" : {
            "description" : "la operation a ete trouver dans la BDD",
            "schema" : {
              "$ref" : "#/definitions/OperationDTO"
            }
          },
          "404" : {
            "description" : "Aucun operation n'existe  dans la BDD avec l'ID fourni"
          }
        }
      }
    },
...
    "/testCovid/v1/patient/{idPatient}" : {
      "get" : {
        "tags" : [ "testCovidv1patient" ],
        "summary" : "rechercher un patient par ID",
        "description" : "Cette methode permet de rechercher  un patient par son ID ",
        "operationId" : "findById",
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "idPatient",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int32"
        } ],
        "responses" : {
          "200" : {
            "description" : "la patient a ete trouver dans la BDD",
            "schema" : {
              "$ref" : "#/definitions/PatientDTO"
            }
          },
          "404" : {
            "description" : "Aucun patient n'existe  dans la BDD avec l'ID fourni"
          }
        }
      }
    },
...
    "/testCovid/v1/preleveur/{idPreleveur}" : {
      "get" : {
        "tags" : [ "preleveur" ],
        "summary" : "rechercher un preleveur par ID",
        "description" : "Cette methode permet de rechercher  un preleveur par son ID ",
        "operationId" : "findById",
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "idPreleveur",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int32"
        } ],
        "responses" : {
          "200" : {
            "description" : "la preleveur a ete trouver dans la BDD",
            "schema" : {
              "$ref" : "#/definitions/PreleveurDTO"
            }
          },
          "404" : {
            "description" : "Aucun preleveur n'existe  dans la BDD avec l'ID fourni"
          }
        }
      }
    },
...
    "/testCovid/v1/recrutement/delete/{idRecrutement}" : {
      "delete" : {
        "tags" : [ "testCovidv1recrutement" ],
        "summary" : "supprimer  un recrutement ",
        "description" : "Cette methode permet de supprimer un recrutement par ID ",
        "operationId" : "delecte",
        "parameters" : [ {
          "name" : "idRecrutement",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int32"
        } ],
        "responses" : {
          "200" : {
            "description" : "recrutement a ete supprimer",
            "schema" : {
              "$ref" : "#/definitions/RecrutementDTO"
            }
          }
        }
      }
    },
...
    "/testCovid/v1/reservation/{idReservation}" : {
      "get" : {
        "tags" : [ "reservation" ],
        "summary" : "rechercher un reservation par ID",
        "description" : "Cette methode permet de rechercher  un reservation par son ID ",
        "operationId" : "findById",
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "idReservation",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int32"
        } ],
        "responses" : {
          "200" : {
            "description" : "la reservation a ete trouver dans la BDD",
            "schema" : {
              "$ref" : "#/definitions/ReservationDTO"
            }
          },
          "404" : {
            "description" : "Aucun reservation n'existe  dans la BDD avec l'ID fourni"
          }
        }
      }
    },
...
    "/testCovid/v1/role/delete/{idRole}" : {
      "delete" : {
        "tags" : [ "testCovidv1roles" ],
        "summary" : "supprimer  un patient ",
        "description" : "Cette methode permet de supprimer un role par ID ",
        "operationId" : "delecte",
        "parameters" : [ {
          "name" : "idRole",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int32"
        } ],
        "responses" : {
          "200" : {
            "description" : "patient a ete supprimer",
            "schema" : {
              "$ref" : "#/definitions/RolesDTO"
            }
          }
        }
      }
    },
    "/testCovid/v1/role/{idRole}" : {
      "get" : {
        "tags" : [ "testCovidv1roles" ],
        "summary" : "rechercher un role par ID",
        "description" : "Cette methode permet de rechercher  un role par son ID ",
        "operationId" : "findById",
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "idRole",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int32"
        } ],
        "responses" : {
          "200" : {
            "description" : "la patient a ete trouver dans la BDD",
            "schema" : {
              "$ref" : "#/definitions/RolesDTO"
            }
          },
          "404" : {
            "description" : "Aucun patient n'existe  dans la BDD avec l'ID fourni"
          }
        }
      }
    },
...
    "/testCovid/v1/stock/{idStock}" : {
      "get" : {
        "tags" : [ "stock" ],
        "summary" : "rechercher un stock par ID",
        "description" : "Cette methode permet de rechercher  un stock par son ID ",
        "operationId" : "findById",
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "idStock",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int32"
        } ],
        "responses" : {
          "200" : {
            "description" : "la stock a ete trouver dans la BDD",
            "schema" : {
              "$ref" : "#/definitions/StockDTO"
            }
          },
          "404" : {
            "description" : "Aucun stock n'existe  dans la BDD avec l'ID fourni"
          }
        }
      }
    },
...
    "/testCovid/v1/typeAnalyse/delete/{idTypeAnalyse}" : {
      "delete" : {
        "tags" : [ "typeAnalyse" ],
        "summary" : "supprimer  un typeAnalyse ",
        "description" : "Cette methode permet de supprimer un typeAnalyse par ID ",
        "operationId" : "delecte",
        "parameters" : [ {
          "name" : "idTypeAnalyse",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int32"
        } ],
        "responses" : {
          "200" : {
            "description" : "typeAnalyse a ete supprimer",
            "schema" : {
              "$ref" : "#/definitions/TypeAnalyseDTO"
            }
          }
        }
      }
    },
...
    "/testCovid/v1/typeAnalyse/{idTypeAnalyse}" : {
      "get" : {
        "tags" : [ "typeAnalyse" ],
        "summary" : "rechercher un typeAnalyse par ID",
        "description" : "Cette methode permet de rechercher  un typeAnalyse par son ID ",
        "operationId" : "findById",
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "idTypeAnalyse",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int32"
        } ],
        "responses" : {
          "200" : {
            "description" : "la typeAnalyse a ete trouver dans la BDD",
            "schema" : {
              "$ref" : "#/definitions/TypeAnalyseDTO"
            }
          },
          "404" : {
            "description" : "Aucun typeAnalyse n'existe  dans la BDD avec l'ID fourni"
          }
        }
      }
    }
  },
  "definitions" : {
    "AbsenceDTO" : {
      "type" : "object",
      "properties" : {
        "idAbsence" : {
          "type" : "integer",
          "format" : "int32"
        },
        "dateDebut" : {
          "type" : "string",
          "format" : "date-time"
        },
        "dateFin" : {
          "type" : "string",
          "format" : "date-time"
        },
        "justifie" : {
          "type" : "string"
        },
        "motif" : {
          "type" : "string"
        },
        "pieceJustificatifString" : {
          "type" : "string"
        },
        "employe" : {
          "$ref" : "#/definitions/EmployeDTO"
        }
      }
    },
    "AnalyseMedicalDTO" : {
      "type" : "object",
      "properties" : {
        "idAnalyse" : {
          "type" : "integer",
          "format" : "int32"
        },
        "dateAnalyse" : {
          "type" : "string",
          "format" : "date-time"
        },
        "descriprtion" : {
          "type" : "string"
        },
        "prixAnalyse" : {
          "type" : "number"
        },
        "resultat" : {
          "type" : "boolean"
        },
        "typeAnalyse" : {
          "$ref" : "#/definitions/TypeAnalyseDTO"
        },
        "reservation" : {
          "$ref" : "#/definitions/ReservationDTO"
        }
      }
    },
    "ArticleDTO" : {
      "type" : "object",
      "properties" : {
        "idArticle" : {
          "type" : "integer",
          "format" : "int32"
        },
        "prixArticle" : {
          "type" : "string"
        },
        "typeArticle" : {
          "type" : "string"
        },
        "codeBarre" : {
          "type" : "string"
        },
        "nomArticle" : {
          "type" : "string"
        }
      }
    },
...
    "BilanFinancierDTO" : {
      "type" : "object",
      "properties" : {
        "idBilan" : {
          "type" : "integer",
          "format" : "int32"
        },
        "description" : {
          "type" : "string"
        }
      }
    },
    "BiologisteDTO" : {
      "type" : "object",
      "properties" : {
        "idBiologiste" : {
          "type" : "integer",
          "format" : "int32"
        },
        "nom" : {
          "type" : "string"
        },
        "prenom" : {
          "type" : "string"
        },
        "email" : {
          "type" : "string"
        },
        "sexe" : {
          "type" : "string"
        },
        "telephone" : {
          "type" : "string"
        },
        "codeBiologiste" : {
          "type" : "string"
        },
        "service" : {
          "type" : "string"
        },
        "departement" : {
          "$ref" : "#/definitions/DepartementDTO"
        }
      }
    },
    "BonCommandeDTO" : {
      "type" : "object",
      "properties" : {
        "idBonCommande" : {
          "type" : "integer",
          "format" : "int32"
        },
        "dateCommande" : {
          "type" : "string",
          "format" : "date-time"
        },
        "fournisseur" : {
          "$ref" : "#/definitions/FournisseurDTO"
        },
        "article" : {
          "$ref" : "#/definitions/ArticleDTO"
        },
        "quantiteCommande" : {
          "type" : "integer",
          "format" : "int32"
        }
      }
    },
    "CompteDTO" : {
      "type" : "object",
      "properties" : {
        "idCompte" : {
          "type" : "integer",
          "format" : "int32"
        },
        "libCompte" : {
          "type" : "string"
        },
        "numeroCompte" : {
          "type" : "string"
        }
      }
    },
    "CongeDTO" : {
      "type" : "object",
      "properties" : {
        "idConge" : {
          "type" : "integer",
          "format" : "int32"
        },
        "dateDebut" : {
          "type" : "string",
          "format" : "date-time"
        },
        "dateFin" : {
          "type" : "string",
          "format" : "date-time"
        },
        "employe" : {
          "$ref" : "#/definitions/EmployeDTO"
        }
      }
    },
...
  }
}